+++ drift/src/main/java/drift/com/drift/managers/PresentationManager.java
@@ -139,7 +139,7 @@ public class PresentationManager {
 
     private void showPopupForMessage(final Message message, final int otherMessages) {
 
-        User user = UserManager.getInstance().getUserForId(message.authorId.intValue());
+        User user = UserManager.getInstance().getUserForId(message.authorId);
         Auth auth = Auth.getInstance();
         if (user != null) {
             showPopupForMessage(user, message, otherMessages);
@@ -147,7 +147,7 @@ public class PresentationManager {
             UserManager.getInstance().getUsers(auth.endUser.orgId, new UserManagerCallback() {
                 @Override
                 public void didLoadUsers(Boolean success) {
-                    User user = UserManager.getInstance().getUserForId(message.authorId.intValue());
+                    User user = UserManager.getInstance().getUserForId(message.authorId);
                     showPopupForMessage(user, message, otherMessages);
                 }
             });
